:fire: remove useless code not necessarry for this project
