+++ Makefile
@@ -6,12 +6,12 @@
 ##
 
 # All the source files
-SOURCES			= 	src/execute_command.c \
+SOURCES			= 	src/ftrace.c \
+					src/parse_cmd.c \
+					src/execute_command.c \
 					src/main_loop.c \
-					src/parse_arg.c \
 					src/print_help.c \
 					src/print_syscall_info.c \
-					src/print_syscall_info_detailled.c \
 					src/follow_pid.c
 
 TESTS			=
@@ -101,7 +101,7 @@ $(TESTS_OBJS):	%.o: %.c
 
 tests_run: fclean $(OBJS) $(TESTS_OBJS)
 	@printf "$(RUNNING) $(BLUE) 🔗  Linking$(RESET)"
-	@$(CC) -o tests.out $(filter-out src/main.o, $(OBJS)) \
+	@$(CC) -o tests.out $(filter-out src/ftrace.o, $(OBJS)) \
 	$(TESTS_OBJS) $(FLAGS) -lcriterion >> $(LOG) 2>&1 \
 	&& printf "\r$(SUCCESS)\n" || printf "\r$(FAILURE)\n";
 	@printf "$(RUNNING)$(BLUE)  🧪  Running tests$(RESET)" \
