+++ src/print_help.c
@@ -9,9 +9,6 @@
 
 int print_help(int status)
 {
-    printf("USAGE: ./ftrace [-s] [-p <pid>|<command>]\n");
-    printf("    -s: print detailled output\n");
-    printf("    -p: attach to a running process\n");
-    printf("    <command>: execute the command\n");
+    printf("USAGE: ftrace <command>\n");
     return status;
 }
